% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{consumption_plot}
\alias{consumption_plot}
\title{Plot for consumption}
\usage{
consumption_plot(data, consumption_col = "consumption_0",
  time_col = "end_time")
}
\arguments{
\item{data}{Dataframe with data}

\item{consumption_col}{Name of column with consumption}

\item{time_col}{Name of column with time}
}
\value{
ggplot
}
\description{
Plot for consumption
}
