% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fill_lonely_na}
\alias{fill_lonely_na}
\title{Replace NA value with a previous one.}
\usage{
fill_lonely_na(vec)
}
\arguments{
\item{vec}{Vector of values}
}
\value{
Vector of values with replaced single missing values.
}
\description{
Replacement will happen only if NA value is next to existing values or is the
last value.
}
