% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_month_data}
\alias{filter_month_data}
\title{Filter data from specific month}
\usage{
filter_month_data(data, month_number, time_col = "end_time")
}
\arguments{
\item{data}{Dataset}

\item{month_number}{Number of month (from 1 to 12)}

\item{time_col}{Name of column with time}
}
\value{
Dataset from specific month
}
\description{
Filter data from specific month
}
