% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{median_model}
\alias{median_model}
\title{Get accuracy of model based on median value of consumption}
\usage{
median_model(data, days = 1:7, decision_attr_prefix = "decision_attr_",
  consumption_prefix = "consumption_", num_previous_days = 7)
}
\arguments{
\item{data}{Dataset.}

\item{days}{Number of days for which prediction should be made.}

\item{decision_attr_prefix}{Prefix of columns with decision attribute.}

\item{consumption_prefix}{Prfix of column with consumption.}

\item{num_previous_days}{For how many past days median value should be
calculated?}
}
\value{
RMSE of this model for each horizon.
}
\description{
Get accuracy of model based on median value of consumption
}
