% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{naive_models_report}
\alias{naive_models_report}
\title{Main function for creating report with naive models}
\usage{
naive_models_report(path = "./data/datasets/", pattern = "*",
  read_file_function = readr::read_csv, ...)
}
\arguments{
\item{path}{Directory with datasets.}

\item{pattern}{Regular expression. Only file names which match the
regular expression will be used from directory \code{path}.}

\item{read_file_function}{Function used for reading files from \code{path}.}

\item{...}{Additional parameters for function \code{read_file_function}.}
}
\value{
Information for report with naive models.
}
\description{
Main function for creating report with naive models
}
