% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tables.R
\name{naive_models_summary}
\alias{naive_models_summary}
\title{Prepare table for summary of naive models for report}
\usage{
naive_models_summary(table)
}
\arguments{
\item{table}{Table with summary of errors}
}
\value{
DT datatable
}
\description{
Prepare table for summary of naive models for report
}
