% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_tables.R
\name{naive_models_tables}
\alias{naive_models_tables}
\title{Knitr table for showing vector of values in naive model's report.}
\usage{
naive_models_tables(vector)
}
\arguments{
\item{vector}{Vector of values.}
}
\value{
Knitr kable
}
\description{
In this vector result for each day should be presented.
}
