% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naive_models.R
\name{summary_of_models}
\alias{summary_of_models}
\title{Create summary of models for all datasets}
\usage{
summary_of_models(models)
}
\arguments{
\item{models}{List with results of all models for all datasets.}
}
\value{
Summary.
}
\description{
Create summary of models for all datasets
}
